+++ src/controllers/SectionController.php
@@ -32,7 +32,7 @@ class SectionController extends Controller
         /** @var Section $section */
         $section = $config['tabs'][$segments[1]]['columns'][$segments[2]]['sections'][$segments[3]];
 
-        if (!$section->getPermittedSections('viewentries')) {
+        if ($section->section && !$section->getPermittedSections('viewentries')) {
             throw new ForbiddenHttpException();
         }