+++ src/events/DefineTabsEvent.php
@@ -2,12 +2,14 @@
 
 namespace wsydney76\contentoverview\events;
 
- use Illuminate\Support\Collection;
- use wsydney76\contentoverview\models\Page;
- use yii\base\Event;
+use craft\elements\User;
+use Illuminate\Support\Collection;
+use wsydney76\contentoverview\models\Page;
+use yii\base\Event;
 
- class DefineTabsEvent extends Event
+class DefineTabsEvent extends Event
 {
+    public User $user;
     public Page $page;
     public Collection $tabs;
 }