+++ src/services/ContentOverviewService.php
@@ -4,6 +4,8 @@ namespace wsydney76\contentoverview\services;
 
 use Craft;
 use craft\base\Component;
+use Illuminate\Support\Collection;
+use wsydney76\contentoverview\events\DefinePagesEvent;
 use wsydney76\contentoverview\models\Action;
 use wsydney76\contentoverview\models\Column;
 use wsydney76\contentoverview\models\CustomSection;
@@ -12,10 +14,12 @@ use wsydney76\contentoverview\models\Section;
 use wsydney76\contentoverview\models\Tab;
 use wsydney76\contentoverview\models\WidgetSection;
 use wsydney76\contentoverview\Plugin;
+use function collect;
 
 class ContentOverviewService extends Component
 {
 
+    public const EVENT_DEFINE_PAGES = 'eventDefinePages';
 
     /**
      * @param string $pageKey
@@ -108,6 +112,57 @@ class ContentOverviewService extends Component
     public function createAction(string $className = null): Action
     {
         return Craft::createObject($className ?? Plugin::getInstance()->getSettings()->actionClass);
+    }
+
+    /**
+     * Returns a collection of page configs available for the current user
+     *
+     * @return Collection
+     */
+    public function getPages($isSidebar = false): Collection
+    {
+        $settings = Plugin::getInstance()->getSettings();
+        $pages = $settings->pages;
+        if (!$pages) {
+            // create a single page for use in list widgets
+            $pages = [
+                $this->defaultPage => [
+                    'label' => $settings->pluginTitle,
+                    'url' => 'contentoverview'
+                ]
+            ];
+        }
+
+        // Drop pages if restricted to a user group
+        $currentUser = Craft::$app->user->identity;
+        $pages = collect($pages)
+            ->filter(function($page) use ($currentUser) {
+                return $currentUser->admin || !isset($page['group']) || $currentUser->isInGroup($page['group']);
+            });
+
+        // Drop group heading pseudo pages if links not displayed in sidebar block
+        if (!$isSidebar) {
+            $pages = $pages->filter(function($page) {
+                return !isset($page['heading']);
+            });
+        }
+
+        // Create Page models
+        $pages = $pages->map(fn($page, $key) => $this->createPage($key, $page));
+
+        // Give custom modules the chance to modify pages
+        if ($this->hasEventHandlers(self::EVENT_DEFINE_PAGES)) {
+            $event = new DefinePagesEvent([
+                'user' => Craft::$app->user->identity,
+                'pages' => $pages
+            ]);
+
+            $this->trigger(self::EVENT_DEFINE_PAGES, $event);
+
+            $pages = $event->pages;
+        }
+
+        return $pages;
     }
 
 }